eigen toetsen gemaakt.

+++ LerenTypen/Models/TestTable.cs
@@ -16,7 +16,7 @@ namespace LerenTypen
         public int DifficultyBinder { get; set; }
         public bool IsPrivate { get; set; }
         public int TestId { get; set; }
-        public decimal Rating { get; set; }
+        public double Rating { get; set; }
 
         public TestTable(int number, string name, int testID)
         {
@@ -43,7 +43,7 @@ namespace LerenTypen
         }
 
 
-        public TestTable(int number, string name, int timesMade, int wordHighscore, int amountOfWords, int testDifficulty, string uploader , decimal reviewscore)
+        public TestTable(int number, string name, int timesMade, int wordHighscore, int amountOfWords, int testDifficulty, string uploader , double reviewscore)
         {
             this.WPFNumber = number;
             this.WPFName = name;
@@ -71,7 +71,7 @@ namespace LerenTypen
             }
 
         }
-        public TestTable(int number, string name, int timesMade, int wordHighscore, int amountOfWords, int testDifficulty, string uploader, decimal reviewscore, int isPrivate, int testId)
+        public TestTable(int number, string name, int timesMade, int wordHighscore, int amountOfWords, int testDifficulty, string uploader, double reviewscore, int isPrivate, int testId)
         {
             this.WPFNumber = number;
             this.WPFName = name;
